requirements.txt
paddlepaddle==3.0.0
paddleocr==3.0.2
langchain==0.3.26
langchain-openai==0.3.25
python-dotenv==1.1.1
Pillow==11.2.1
numpy==2.0.2 